+++ packages/icons/src/components/ChevronRightIcon.tsx
@@ -2,13 +2,20 @@ import { Icon, IconProps } from '@chakra-ui/react';
 import React, { FC } from 'react';
 
 const ChevronRightIcon: FC<IconProps> = (props) => (
-  <Icon width={4} height={4} viewBox="0 0 32 32" fill="none" {...props}>
+  <Icon
+    width={4}
+    height={4}
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    viewBox="0 0 24 24"
+    strokeWidth={1.5}
+    stroke="currentColor"
+    {...props}
+  >
     <path
-      d="M7.5 15L12.5 10L7.5 5"
-      stroke="currentColor"
-      strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
+      d="M8.25 4.5l7.5 7.5-7.5 7.5"
     />
   </Icon>
 );
