+++ packages/icons/src/components/company/HovIcon.tsx
