+++ packages/icons/src/components/social/TwitterIcon.tsx
@@ -1,15 +1,18 @@
-import { Icon } from '@chakra-ui/react';
+import { Icon, IconProps } from '@chakra-ui/react';
 import React, { FC } from 'react';
-export interface TwitterIconProps {
+
+type BaseStyle = Omit<IconProps, 'isDisabled' | 'variant'>;
+
+export interface TwitterIconProps extends BaseStyle {
   isDisabled?: boolean;
   variant?: 'outline' | 'solid';
 }
 
 const TwitterIcon: FC<TwitterIconProps> = (props) => {
-  const { isDisabled, variant = 'outline' } = props;
+  const { isDisabled, variant = 'outline', ...others } = props;
 
   return (
-    <Icon width="24px" verticalAlign={'center'} viewBox="0 0 25 20" fill="none">
+    <Icon width={4} height={4} viewBox="0 0 25 20" fill="none" {...others}>
       <path
         d="M7.55016 19.75C16.6045 19.75 21.5583 12.2467 21.5583 5.74186C21.5583 5.53092 21.5536 5.3153 21.5442 5.10436C22.5079 4.40746 23.3395 3.54425 24 2.5553C23.1025 2.9546 22.1496 3.21538 21.1739 3.32874C22.2013 2.71291 22.9705 1.74547 23.3391 0.605767C22.3726 1.17856 21.3156 1.58261 20.2134 1.80061C19.4708 1.01156 18.489 0.489116 17.4197 0.314051C16.3504 0.138986 15.2532 0.32105 14.2977 0.832096C13.3423 1.34314 12.5818 2.15471 12.1338 3.14131C11.6859 4.12792 11.5754 5.23462 11.8195 6.2903C9.86249 6.19209 7.94794 5.6837 6.19998 4.7981C4.45203 3.91249 2.90969 2.66944 1.67297 1.14952C1.0444 2.23324 0.852057 3.51565 1.13503 4.73609C1.418 5.95654 2.15506 7.02345 3.19641 7.71999C2.41463 7.69517 1.64998 7.48468 0.965625 7.10592V7.16686C0.964925 8.30415 1.3581 9.40659 2.07831 10.2868C2.79852 11.167 3.80132 11.7706 4.91625 11.995C4.19206 12.1931 3.43198 12.222 2.69484 12.0794C3.00945 13.0574 3.62157 13.9129 4.44577 14.5263C5.26997 15.1398 6.26512 15.4806 7.29234 15.5012C5.54842 16.8711 3.39417 17.6141 1.17656 17.6106C0.783287 17.61 0.390399 17.5859 0 17.5384C2.25286 18.9837 4.87353 19.7514 7.55016 19.75Z"
         fill={
