+++ packages/icons/src/index.tsx
@@ -5,13 +5,13 @@ export { default as CheckCircleIcon } from './components/CheckCircleIcon';
 export { default as CheckIcon } from './components/CheckIcon';
 export { default as ChevronLeftIcon } from './components/ChevronLeftIcon';
 export { default as ChevronRightIcon } from './components/ChevronRightIcon';
+export { default as HovIcon } from './components/company/HovIcon';
 export { default as DotsVerticalIcon } from './components/DotsVerticalIcon';
 export { default as ErrorFolderIcon } from './components/ErrorFolderIcon';
 export { default as ExitIcon } from './components/ExitIcon';
 export { default as HelpCircleIcon } from './components/HelpCircleIcon';
 export { default as HelpIcon } from './components/HelpIcon';
 export { default as HomeIcon } from './components/HomeIcon';
-export { default as HovIcon } from './components/HovIcon';
 export { default as LayersTwoIcon } from './components/LayersTwoIcon';
 export { default as MessageSmileIcon } from './components/MessageSmileIcon';
 export { default as PencilIcon } from './components/PencilIcon';
