+++ packages/icons/src/story/Icons.stories.tsx
@@ -5,11 +5,11 @@ import CheckCircleIcon from '../components/CheckCircleIcon';
 import CheckIcon from '../components/CheckIcon';
 import ChevronLeftIcon from '../components/ChevronLeftIcon';
 import ChevronRightIcon from '../components/ChevronRightIcon';
+import HovIcon from '../components/company/HovIcon';
 import ErrorFolderIcon from '../components/ErrorFolderIcon';
 import ExitIcon from '../components/ExitIcon';
 import HelpIcon from '../components/HelpIcon';
 import HomeIcon from '../components/HomeIcon';
-import HovIcon from '../components/HovIcon';
 import LayersTwoIcon from '../components/LayersTwoIcon';
 import MessageSmileIcon from '../components/MessageSmileIcon';
 import PrimaryIcon from '../components/PrimaryIcon';
